Move getTransformFlip from design-system to element-library
