+++ packages/element-library/src/index.js
@@ -18,5 +18,6 @@ export * from './shared';
 export * from './media';
 export * from './utils/textMeasurements';
 export * from './text/util';
+export { default as getTransformFlip } from './utils/getTransformFlip';
 export { default as TextOutput, TextOutputWithUnits } from './text/output';
 export { default as elementTypes } from './elementTypes';
