+++ packages/element-library/src/media/editCropMoveable.js
@@ -22,8 +22,12 @@ import { useEffect, useRef } from '@googleforcreators/react';
 import { useUnits, calcRotatedResizeOffset } from '@googleforcreators/units';
 import { getFocalFromOffset } from '@googleforcreators/media';
 import { Moveable } from '@googleforcreators/moveable';
-import { getTransformFlip } from '@googleforcreators/design-system';
 import { StoryPropTypes } from '@googleforcreators/elements';
+
+/**
+ * Internal dependencies
+ */
+import getTransformFlip from '../utils/getTransformFlip';
 
 function EditCropMoveable({
   setProperties,
