+++ packages/element-library/src/video/trim.js
@@ -21,10 +21,7 @@ import styled from 'styled-components';
 import PropTypes from 'prop-types';
 import { useRef, useMemo, useCallback } from '@googleforcreators/react';
 import { getMediaSizePositionProps } from '@googleforcreators/media';
-import {
-  CircularProgress,
-  getTransformFlip,
-} from '@googleforcreators/design-system';
+import { CircularProgress } from '@googleforcreators/design-system';
 import { StoryPropTypes } from '@googleforcreators/elements';
 
 /**
@@ -32,6 +29,7 @@ import { StoryPropTypes } from '@googleforcreators/elements';
  */
 import MediaDisplay from '../media/display';
 import { elementWithFlip } from '../shared';
+import getTransformFlip from '../utils/getTransformFlip';
 import PlayPauseButton from './playPauseButton';
 import { getBackgroundStyle, videoWithScale } from './util';
 import Captions from './captions';
