+++ packages/masks/package.json
@@ -37,6 +37,7 @@
   },
   "dependencies": {
     "@googleforcreators/design-system": "*",
+    "@googleforcreators/element-library": "*",
     "@googleforcreators/elements": "*",
     "@googleforcreators/i18n": "*",
     "@googleforcreators/react": "*",
